Remove pyinstxtractor.py

+++ make.py
@@ -2,7 +2,6 @@ import importlib.util
 import logging
 import os
 import shutil
-import sys
 from subprocess import Popen, PIPE
 
 import PyInstaller.__main__
@@ -11,10 +10,9 @@ from PyInstaller.utils.hooks import collect_submodules
 
 from IceSpringMusicPlayer.utils.logUtils import LogUtils
 
+name = "IceSpringMusicPlayer"
 LogUtils.initLogging()
 logging.getLogger().setLevel(logging.INFO)
-
-name = "IceSpringMusicPlayer"
 logging.info("Removing application directory if exists...")
 Path(f"dist/{name}").rmtree(ignore_errors=True)
 
@@ -27,28 +25,46 @@ else:
     imports = collect_submodules('IceSpringMusicPlayer') + ["statistics", "scipy.signal"]
     hiddenImports = [("--hidden-import", x) for x in imports]
     hiddenImports = [y for x in hiddenImports for y in x]
-    PyInstaller.__main__.run(["main.py", "--noconsole", "--noupx", *hiddenImports, "--name", name])
+    PyInstaller.__main__.run(["main.py", "--noupx", "-d", "noarchive", *hiddenImports, "--name", name])
     Path(f"dist/{name}").copytree(Path(f"dist/{name}-backup"))
 
 logging.info("Change directory to application root")
 os.chdir(f"dist/{name}")
-logging.info("Extract pyinstaller package...")
-os.system(f"{sys.executable} ../../resources/pyinstxtractor.py {name}.exe")
 
-logging.info("Copying 3rd libraries to pylib...")
-root = Path(f"{name}.exe_extracted/PYZ-00.pyz_extracted")
-for path in Path(root).glob("**/*.pyc"):
-    if path.relative_to(root).parts[0] == name:
-        continue
-    module = ".".join(x.split(".pyc")[0] for x in path.relative_to(root).parts if x != "__init__.pyc")
-    origin = importlib.util.find_spec(module, None).origin
-    if not module == "importlib._bootstrap":
-        assert origin is not None
-    origin = Path(origin) if origin not in [None, "frozen"] else path
-    target = Path("pylib") / path.relative_to(root).with_suffix(origin.suffix)
-    logging.debug("Copying %s => %s", origin, target)
-    target.parent.mkdir(parents=True, exist_ok=True)
-    shutil.copyfile(origin, target)
+logging.info("Moving python dependencies to pylib...")
+for path in Path().glob("*"):
+    if path.is_dir() or path.suffix in (".pyc", ".pyd"):
+        target = Path("pylib") / path
+        logging.debug("Moving %s => %s", path.absolute(), target.absolute())
+        target.parent.mkdir(parents=True, exist_ok=True)
+        if not target.exists():
+            shutil.move(str(path), str(target))
+
+logging.info("Removing application compiled files")
+shutil.rmtree(f"pylib/{name}")
+
+removingPyc = True
+logging.info("Remove pyc enabled: %s", removingPyc)
+if removingPyc:
+    logging.info("Changing .pyc/.pyo to .py...")
+    count = 0
+    for path in Path("pylib").glob("**/*"):
+        if path.suffix not in (".pyc", ".pyo"):
+            continue
+        if "__pycache__" in path.parts:
+            continue
+        module = ".".join(x for x in path.relative_to("pylib").with_suffix("").parts if x != "__init__")
+        origin = importlib.util.find_spec(module).origin
+        if origin is not None and Path(origin).exists() and Path(origin).suffix == ".py":
+            count += 1
+            path.unlink()
+            logging.debug("Copying %d %s => %s", count, origin, path.with_suffix(".py"))
+            shutil.copyfile(origin, path.with_suffix(".py"))
+    logging.info("%d .pyc/.pyo files processed", count)
+
+logging.info("Copying pydub source files...")
+shutil.copyfile(importlib.util.find_spec("pydub.utils").origin, "pylib/pydub/utils.py")
+shutil.copyfile(importlib.util.find_spec("pydub.audio_segment").origin, "pylib/pydub/audio_segment.py")
 
 logging.info("Copying application source files...")
 for path in Path(f"../../{name}").glob("**/*.py"):
@@ -57,27 +73,6 @@ for path in Path(f"../../{name}").glob("**/*.py"):
     target.parent.mkdir(parents=True, exist_ok=True)
     shutil.copyfile(path, target)
 shutil.copyfile("../../main.py", "main.py")
-
-logging.info("Moving python dynamic libraries to pydll...")
-for path in Path().glob("*"):
-    if path.name.endswith(".exe_extracted"):
-        continue
-    if path.name in ("pylib", "pydll", name):
-        continue
-    if path.is_dir() or path.suffix == ".pyd":
-        target = Path("pydll") / path
-        logging.debug("Moving %s => %s", path.absolute(), target.absolute())
-        target.parent.mkdir(parents=True, exist_ok=True)
-        if not target.exists():
-            shutil.move(str(path), str(target))
-
-logging.info("Merging pydll to pylib...")
-for path in Path("pydll").glob("**/*"):
-    target = Path("pylib") / path.relative_to("pydll")
-    logging.debug("Copying %s => %s", path.absolute(), target.absolute())
-    if path.is_file() and not target.exists():
-        target.parent.mkdir(parents=True, exist_ok=True)
-        shutil.copy(path, target)
 
 logging.info("Copying resources....")
 for path in Path("../../resources").glob("**/*"):
@@ -100,7 +95,7 @@ cd ../../resources
 "{vs}"\BuildTools\VC\Auxiliary\Build\vcvarsall.bat x86_amd64
 rc.exe resources.rc
 cl.exe /c main.c -I {python}\include
-cl.exe main.obj /link /LIBPATH {python}\libs\python37.lib resources.res {flag}
+cl.exe main.obj resources.res /link /LIBPATH {python}\libs\python37.lib {flag}
 del main.obj resources.res
 """.strip().splitlines()
 proc = Popen("cmd", stdin=PIPE, stdout=PIPE, stderr=PIPE)
@@ -117,14 +112,11 @@ shutil.move("../../resources/main.exe", f"{name}.exe")
 logging.info("Cleaning...")
 excluded_files = "Qt5DataVisualization.dll Qt5Pdf.dll Qt5Quick.dll Qt5VirtualKeyboard.dll d3dcompiler_47.dll " \
                  f"libGLESv2.dll opengl32sw.dll base_library.zip {name}.exe_extracted pydll".split()
-for file in Path().glob("*"):
-    if file.name in excluded_files:
-        if file.is_dir():
-            logging.info(f"Removing folder %s...", file.name)
-            shutil.rmtree(file)
-        else:
-            logging.info(f"Removing file %s...", file.name)
-            file.unlink()
+for filename in excluded_files:
+    file = Path(filename)
+    if file.exists():
+        logging.info(f"Removing %s...", file.name)
+        shutil.rmtree(file) if file.is_dir() else file.unlink()
 
-Path(f"{name}.7z").unlink(missing_ok=True)
+Path(f"../../{name}.7z").unlink(missing_ok=True)
 os.system(f"cd ../../dist && 7z a -mx=9 ../{name}.7z {name}")
